% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputfile.R
\name{generarOutput}
\alias{generarOutput}
\title{Generating Output}
\usage{
generarOutput(output, config, path)
}
\arguments{
\item{output}{}

\item{config}{}

\item{path}{}
}
\value{

}
\description{
generate output file with the user's request
}
\author{
MigBor
}
