% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(df_merge_final, config)
}
\arguments{
\item{config}{}
}
\value{
output
}
\description{
search for the required output
}
\author{
Fran&Cyn
}
