% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{regression_model}
\alias{regression_model}
\title{regression model}
\usage{
regression_model(df_merge_final, config)
}
\arguments{
\item{config}{}
}
\value{
final_prediction
}
\description{
Code to run the linear regression model.
}
\author{
Fran&Cyn
}
